5.1.5
Auto-done added. If all items in cart, list is marked as Done. If an
item is then removed, the list is marked as Not Done. Can still override
in ListUpdateView.
Removed two exception catch blocks, replaced with get_or_404.
List.modified no longer shown in listdetailview, was found to be
unneeded.
List edit button made bigger.